currentGrowth increases on its own using SimpleTimer, all thirst is converted into currentGrowth and increases it, currentGrowth stops increasing on its own when maxThirst reached

+++ PlantData.java
@@ -17,7 +17,7 @@ public class PlantData implements Serializable
     
     public int age = 0; // age is number of waters
     public int thirst = 0; // num seconds since lastWateredTime
-    public int currentGrowth = 0; // num seconds of growth
+    public double currentGrowth = 0.0; // num seconds of growth
     public int growthStage = 0; // growthStage determines image to use for plant, as well as amount of y to adjust from pot
     
     // Plant species
@@ -34,7 +34,7 @@ public class PlantData implements Serializable
     
     public boolean isThirsty() {
        return thirst >= 15;
-   }
+       }
    
-   
+
 }
